% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pushpin.R
\name{geom_pushpin}
\alias{geom_pushpin}
\title{Pushpin layer}
\usage{
geom_pushpin(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\description{
The geom is used to add pushpins to plots. See ?ggplot2::geom_points for more info. The pushpin indicates the exact point on the pin point. Use the argument \code{pushpin} to select the color. Available options are \code{"red"} (default), \code{"blue"}, \code{"green"}, \code{"yellow"}, \code{"orange"}, \code{"gray"}, \code{"white"} and \code{"transparent"}. Change the size of the puspin with \code{size} as in \code{ggplot2::geom_point}
}
\examples{

# install.packages("ggplot2")
library(ggplot2)


df <- data.frame(x = state.center$x, y = state.center$y,
                state = state.name)

ggplot(df, aes(x = x, y = y)) +
 geom_polygon(data = map_data("state"),
              color = "black",
              fill = "white",
              aes(x = long, y = lat,
                  group = group)) +
 geom_pushpin(pushpin = "blue", size = 1) +
 guides(fill = FALSE) +
 theme_void()

}
